Code refactoring and fixing

+++ src/Helpers/Dir.php
@@ -4,35 +4,31 @@ namespace Pabilsag\Helpers\Dir;
 
 use function Pabilsag\Helpers\Path\standardize_file_path;
 
-function rmdir_recursive (string $path): void
+function rmdir_recursive(string $path): void
 {
-	$cleaned_path = standardize_file_path($path);
+    $path = standardize_file_path($path);
 
-    $list = array_diff(scandir($cleaned_path), [ '.', '..' ]);
+    if(!is_dir($path))
+	{
+        return;
+    }
 
-    try
-    {
-        foreach($list as $file):
+    $items = array_diff(scandir($path), ['.', '..']);
 
-            $filepath = "{$cleaned_path}/{$file}";
+    foreach($items as $item)
+	{
+        $filepath = $path . '/' . $item;
 
-            if(is_dir($filepath))
-            {
-                rmdir_recursive($filepath);
-                continue;
-            }
-
+        if(is_dir($filepath))
+		{
+            rmdir_recursive($filepath);
+        }
+		else
+		{
             unlink($filepath);
-
-        endforeach;
-
-		// Remove path after it is completely empty
-	    rmdir($path);
+        }
     }
-	catch(\Throwable $err)
-    {
-		throw new \Exception($err->getMessage());
-    }
+
+    rmdir($path);
 }
 
-?>
